Sunburst chart (#4037)
## Description

Adds an entirely new chart type, the Sunburst Chart - resolves #3842. 

Note: this is a WIP. I've added a Typescript component for a responsive
sunburst chart, a story to demonstrate the functionality, and mock data.
I still need to add event handlers, and a tooltip - I wanted to give
this a chance to be reviewed before changes grow too large, and I'm
looking for guidance on the best way to integrate these elements to
match the style of existing components.

## Related Issue

#3842 

## Motivation and Context

This adds a popular and useful new hierarchical chart type that is not
addressed by any existing component in the library.

## How Has This Been Tested?

I have not added tests yet. 

## Screenshots (if appropriate):

Current state of the chart - responsive dimensions and labels, still
needs interactivity.

<img width="660" alt="Screenshot 2024-01-03 at 6 59 24 PM"
src="https://github.com/recharts/recharts/assets/59377951/6fbad951-f864-42c0-9c55-d70a4bf76b6e">

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] My code follows the code style of this project.
- [X] My change requires a change to the documentation.
- [X] I have updated the documentation accordingly.
- [X] I have added tests to cover my changes.
- [X] I have added a storybook story or extended an existing story to
show my changes
- [X] All new and existing tests passed.