Enable developers to specify tooltip position before users interact (#4070)
## Description

Added a `defaultIndex` prop to <strike>all chart types</strike> EDIT:
the `<Tooltip />` element. If someone specifies `defaultIndex={2}`, the
2nd data item will show a tooltip. This includes, when relevant:
- The relevant data will show its hover state
- The vertical line / crosshair appears around the data point
- The floating tooltip label will appear in an appropriate location

Additionally, this works in tandem with the accessibilityManager. If a
developer provides the prop `defaultIndex={2}`, and a keyboard-only user
tabs to that chart, they will jump to the starting position. If the user
presses the right arrow, they will move to index 3.

A story has been added under the storybook "Tooltip" section to
demonstrate this functionality

Other behavioral notes:
- Supports vertical charts
- Setting `defaultIndex` doesn't tamper with the tooltip's usual
interactions. Users can still move around the chart, and the tooltip
will move with them.
- If the `defaultIndex` prop is updated, the chart handles the update as
expected. (Hence, changing the `componentDidUpdate` lifecycle method in
the code.
- If the chart is in a ResponsiveContainer, and the user hasn't
interacted with the chart, and the chart gets resized... then the
tooltip stays in the correct location.
- Extra handling for scatter plots, since they're weird. (They require a
crosshair, which means that we need to know the y-value of the relevant
point, but... we don't have that information in the props or state... so
I dig it out of the graphicalItem associated with the scatter plot.
Don't love this solution, but it's the only way I could find to support
scatter plots for this.)

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->
https://github.com/recharts/recharts/issues/3962

## Motivation and Context

In my library, Chart2Music, we provide extra accessibility support to
data visualizations by turning charts into music. As part of that
process, we provide screen reader users the ability to access and
interact with data in a chart, and visually synchronize the chart to the
user's actions.

While we can provide visual syncing with other charting libraries, such
as react-chartjs-2, it hasn't been available for recharts.

With this change, we'll be able to improve the UX for users with a
Chart2Music+Recharts tech stack.

## How Has This Been Tested?

Added automated tests cover every line that was changed in this PR.

To identify edge cases, I made sure the property would work as expected
on every example in storybook.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] I have added a storybook story or extended an existing story to
show my changes
- [x] All new and existing tests passed.