Tooltip: Enable showing hidden graphicalItems in Tooltip (#4083)
<!--- Provide a general summary of your changes in the Title above -->

## Description
Instead of not passing on the payload of hidden items to the Tooltip at
all, we extend the TooltipPayload with the hide field, and filter before
rendering in the Tooltip.

This solution is better because it easily enables users to show more
data in the Tooltip without rendering it.

It is also is a targeted change, moving more decision and responsibility
into the Tooltip component.

<!--- Describe your changes in detail -->

## Related Issue
https://github.com/recharts/recharts/issues/4078
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
The Tooltip only shows content of element that are visible. 
If we want to influence the chart, but not pollute it with additional
lines / visual drawings, we can hide elements.

Hidden elements already can affect the yAxis or xAxis. 
Now they can also be shown in a Tooltip. 


<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Storybook and Unit Tests

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):
Although only a single line is drawn, the Tooltip contains information
about two values in the payload.
<img width="407" alt="Screenshot 2024-01-26 at 22 48 21"
src="https://github.com/recharts/recharts/assets/27995606/30a6c94a-ea2e-4fa0-88d1-fc88b9d19f63">

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] I have added a storybook story or extended an existing story to
show my changes
- [x] All new and existing tests passed.

---------

Co-authored-by: “Nikolas <“nikolas@rieble.com“>