type improvement (#4082)
<!--- Provide a general summary of your changes in the Title above -->

When using custom labels, its become a sore to have to
determine/remember what props the label function provides.
I'd love it if I can call the label function and know exactly what
properties exist. But currently we just see it typed as 'any' which
doesn't really make sense to me.

## Description

I've separated the properties defined in `PieSectorDataItem` into their
own type called `PieSectorData` for more flexible usage.
I've also defined a new type called `PieLabelProps` to be used as the
extension for `P` in `type PieLabel<P>`.
The new type includes at the minimum, what we already know is returned
in PieLabel props.
Since we utilize `GeometrySector` to help define `PieLabelProps` we may
see some keys defined as optional that won't actually exist in
PieLabelProps - this is because `GeometrySector` was not defined for
this purpose. This is an imperfect solution.
Finally, we add `PieLabelRenderProps` which contains additional
necessary items as well as `[key: string]: any` in its definition,
ensuring there are no breaking changes.

<!--- Describe your changes in detail -->

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [ ] I have added a storybook story or extended an existing story to
show my changes
- [x] All new and existing tests passed.