fix: reference line should not render if its coord is NaN (#4244)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->
Currently when a ReferenceLine is created and the scale generates `NaN`
as a coordinate of the line (X or Y), the line still renders. It renders
with a value of `NaN` which is invalid and should never be in the DOM.
This gets treated as 0 and is rendered on the far edge of the chart.

![image](https://github.com/recharts/recharts/assets/25180830/d3d19f2e-6a83-4b6e-9ffe-92a1bca8393e)

If the main coord of the ReferenceLine is `NaN` we should prefer to
discard it (at least for now) rather than rendering something that
shouldn't be in the DOM.

As a follow up we could warn when this happens or for the case of
duplicated categories, we could attempt to render lines at both
categories somehow.

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->
https://github.com/recharts/recharts/issues/4239

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
Attempts to mitigate some confusing behavior

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->
- unit tests
- manual tests

## Screenshots (if appropriate):
see above

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Breaking change (fix or feature that would cause existing
functionality to change)
  -  technically breaking in terms of current functionality

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [X] My code follows the code style of this project.
- [ ] My change requires a change to the documentation. --> the
documentation currently states that it is discarded but it is not
- [ ] I have updated the documentation accordingly.
- [X] I have added tests to cover my changes.
- [ ] I have added a storybook story or extended an existing story to
show my changes
- [X] All new and existing tests passed.

---------

Co-authored-by: Coltin Kifer <ckifer@amazon.com>