BREAKING CHANGE: Make `react-is` a peer dependency (#4541)
<!--- Provide a general summary of your changes in the Title above -->

## Description

Make `react-is` a peer dependency instead of a direct dependency

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context

`react-is` needs to be the same version as the used version of React. So
far these dependencies could be out of sync since the related internals
didn't changed. However, in React 19 related internals did change. Using
`react-is<19` would not recognize elements from `react>=19` and vice
versa. So generally, `react-is` should be installed by library users
just like React. That way, users can make sure the versions match
without having to resort to `resolutions` or `overrides` which wouldn't
even work for libraries.

## How Has This Been Tested?



## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] I have added a storybook story or extended an existing story to
show my changes