Add axisMapSlice; attempt to write more accurate Axis types (#4585)
## Description

I started adding the new slice and then got sidetracked trying to decide
which type it should be. Anyway. This has green build so good to go I
think. I will continue on the state later.

Recharts has two different Axis props:
- the external one, that our users are seeing,
- and the internal, computed one

Most of all properties are the same in both, but some are added and
what's worse, some exist in both with same name but different type.

Honorable mention for `scale` which exists in three different variants
(SVG attribute, external prop, internal d3 object) but the types were
pretending it's all the same `any`.

## Related Issue

https://github.com/recharts/recharts/issues/4583

## Motivation and Context

The current axisMap is nice but it reads everything from DOM (and then
it sends to Redux). To complete the Tooltip transition we first have to
complete the Axis transition.

## How Has This Been Tested?

npm test

## Screenshots (if appropriate):

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [ ] I have added a storybook story or extended an existing story to
show my changes