CartesianGrid uses redux axes instead of generator axes (#4795)
## Description

Removing another dependency on generator axisMap.

This adds new props, `xAxisId` and `yAxisId` to CartesianGrid. Before,
it used to select "first" available axis. Trouble is, since the axes are
stored in an object, there is no definition of what "first" means, and
different browsers and environments are allowed to iterate object values
in different ways. I very much don't like that so let's add an explicit
axis IDs, and default it to 0, same as all other components do.

## Related Issue

https://github.com/recharts/recharts/issues/4583