Add more stories for RadialBarChart (#5161)
Note that there stories are not very nice charts - they are all broken
in a way. The axis ticks overlap, the data don't render at all, the
sectors cover each other, the axis direction and data direction do not
match. I am adding these because I am planning to introduce some
breaking changes and I want to demonstrate that the changes are making
RadialBarChart better.

I also recreated the same stories in codesandbox to observe the
behaviour in 2.x release:
https://codesandbox.io/p/sandbox/lord-of-the-radial-bar-chart-5h8mlf?file=%2Fsrc%2FApp.tsx%3A166%2C3
you will notice that the last chart axis has the correct orientation in
2.x but wrong orientation in the 3.x branch, I will fix that too.

## Related Issue

https://github.com/recharts/recharts/issues/4583