Proposing new prop for Tooltip (#5237)
## Description

This is only the prop - no implementation yet. I want to have an
agreement on what the external API should look like before implementing.

This option is what I think is the simplest - we let Tooltip stay sticky
to the "categorical axis" and do not let users to override it (this is
the same in 2.x). If in the future we do decide to allow overriding it,
we can add a new prop that goes `axisType: xAxis | yAxis | radiusAxis |
angleAxis`.

What's new is we will allow users to specify which of the categorical
axes exactly should be the tooltip axis - in case there are multiple.

WDYT?

Alternatively we could have a bunch of props that goes like `xAxisId:
AxisId; yAxisId: AxisId; radiusAxisId: AxisId; angleAxisId: AxisId;` but
I selected against it because it will allow users to specify multiple
conflicting axes at the same time. But now that I think of it, that's
exactly what Scatter is doing.

## Related Issue

https://github.com/recharts/recharts/issues/4583