Replace lodash functions with native code (#5268)
## Description
When bundling an application, lodash seems to be taking a lot of space
when using recharts. This is the first PR to replace lodash with either
native functions or move to es-toolkit.

This PR replaces any functions that could be replaced with either
`typeof`, `isNullOrUndefined`, `upperFirst`, and `Number.isNaN`. In the
future, there will be other PRs that replace other instances of lodash
functions that are more complex.

## Related Issue

Issue: https://github.com/recharts/recharts/issues/5266

## Motivation and Context

Reduces the size of recharts bundle by removing lodash and replacing
with internal functions.

## How Has This Been Tested?

All tests pass, and new tests have been written for `isNullOrUndefined`,
and `upperFirst` functions.

## Types of changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [ ] I have added a storybook story or extended an existing story to
show my changes