Backport Line renderDots to include points object in the dotProps (#5657)
<!--- Provide a general summary of your changes in the Title above -->

## Description

`points` was an object accessible through the Line dots object, and I am
using that object to build a graph. That was some versions ago, and I
just updated my recharts version and `points` is no longer there, though
you just put it back 2 days ago in the 3.x branch.

The issue is the same as the Area fix in
https://github.com/recharts/recharts/pull/4481/files?diff=split&w=1,
https://github.com/recharts/recharts/issues/4480,
https://github.com/recharts/recharts/blob/3.x/src/cartesian/Line.tsx#L257,
and https://github.com/recharts/recharts/discussions/5656.

## Related Issue

The issue is the same as the Area fix in
https://github.com/recharts/recharts/pull/4481/files?diff=split&w=1,
https://github.com/recharts/recharts/issues/4480,
https://github.com/recharts/recharts/blob/3.x/src/cartesian/Line.tsx#L257,
and https://github.com/recharts/recharts/discussions/5656.

## Motivation and Context

I am using the points object to build a custom graph, I am way behind in
the recharts version, and I want to bring my code up to the latest
version.

## How Has This Been Tested?

I added almost same test you added for area, and the test passes, not
the pipeline because of a different file that fails with and without my
change.

<img width="1070" alt="Screenshot 2025-03-10 at 3 32 17 PM"
src="https://github.com/user-attachments/assets/48694a65-9fc1-47b5-88b5-b048fcd108df"
/>


## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [ ] I have added a storybook story or extended an existing story to
show my changes

---------

Co-authored-by: Bety E. Rodriguez-Milla <brodriguezmilla@auvik.com>