Don't set dataEndIndex to -1 so that Scatter can show tooltip again (#5716)
## Description

This code was setting the dataEndIndex to `-1` which wasn't detected as
"empty" because we only detect "empty" as "equals to zero" and -1 is not
equal to zero.

This should fix the Scatter tooltip problem that @jurij reported here:
https://github.com/recharts/recharts/issues/5445#issuecomment-2746356448

This bug only happened on when the `<ChartDataContextProvider />` got
unmounted which doesn't happen in the storybook but happens outside and
I am not entirely sure why but we need to guard against that anyway
because React can mount and unmount things as it pleases.

## Related Issue

https://github.com/recharts/recharts/issues/5445